+++ pcs/usage.py
@@ -1522,7 +1522,7 @@ Commands:
             ip_version, knet_pmtud_interval, link_mode
         Link options are:
             link_priority, linknumber, mcastport, ping_interval,
-            ping_precision, ping_timeout, pong_count, transport (udp or sctp)
+            ping_precision, ping_timeout, pong_count, transport
             Each 'link' followed by options sets options for one link in the
             order the links are defined by nodes' addresses. You can set link
             options for a subset of links using a linknumber. See examples
@@ -1592,12 +1592,12 @@ Commands:
                 node1 addr=10.0.1.11 addr=10.0.2.11 \\
                 node2 addr=10.0.1.12 addr=10.0.2.12
         Set link options for all links. Link options are matched to the links
-            in order. The first link (link 0) has sctp transport, the second
-            link (link 1) has mcastport 55405:
+            in order. The first link (link 0) has pong_count 4, the second link
+            (link 1) has mcastport 55405:
             pcs cluster setup newcluster \\
                 node1 addr=10.0.1.11 addr=10.0.2.11 \\
                 node2 addr=10.0.1.12 addr=10.0.2.12 \\
-                transport knet link transport=sctp link mcastport=55405
+                transport knet link pong_count=4 link mcastport=55405
         Set link options for the second and fourth links only. Link options are
             matched to the links based on the linknumber option (the first link
             is link 0):
@@ -1608,7 +1608,7 @@ Commands:
                 addr=10.0.1.12 addr=10.0.2.12 addr=10.0.3.12 addr=10.0.4.12 \\
                 transport knet \\
                 link linknumber=3 mcastport=55405 \\
-                link linknumber=1 transport=sctp
+                link linknumber=1 pong_count=4
         Create a cluster using udp transport with a non-default port:
             pcs cluster setup newcluster node1 node2 \\
                 transport udp link mcastport=55405
@@ -1909,7 +1909,7 @@ Commands:
         linknumber.
         Link options (documented in corosync.conf(5) man page) are:
             link_priority, linknumber, mcastport, ping_interval,
-            ping_precision, ping_timeout, pong_count, transport (udp or sctp)
+            ping_precision, ping_timeout, pong_count, transport
 
     link delete <linknumber> [<linknumber>]...
         Remove specified corosync links.
@@ -1927,7 +1927,7 @@ Commands:
         Link options (documented in corosync.conf(5) man page) are:
         for knet transport:
             link_priority, mcastport, ping_interval, ping_precision,
-            ping_timeout, pong_count, transport (udp or sctp)
+            ping_timeout, pong_count, transport
         for udp and udpu transports:
             bindnetaddr, broadcast, mcastaddr, mcastport, ttl